refactor(gold_standard): move logic to refine gold standards to `L2GGoldStandard`
